pyproject.toml
[project]
name = "articlegen"
version = "0.1.0"
description = "A flask project which is used as api endpoint for generating blog response and provide to the frontend."
authors = [
    { name = "Jainam Sanghavi", email = "jainamsanghavi008@gmail.com" }
]
readme = "README.md"

requires-python = ">=3.12,<4.0"


dependencies = [
    "aiohappyeyeballs==2.4.6",
    "aiohttp==3.11.13",
    "attrs==25.1.0",
    "Flask==3.1.0",
    "flask-cors==5.0.1",
    "gunicorn==23.0.0",
    "numpy==2.2.3",
    "SQLAlchemy==2.0.38",
    "requests==2.32.3",
    "langchain==0.3.19",
    "langchain-core==0.3.39",
    "langchain-groq==0.2.4",
    "pydantic==2.10.6",
    "python-dotenv==1.0.1"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
